refactor test data to text-fixtures crate
