+++ protocol/test-fixtures/src/lib.rs
@@ -40,6 +40,8 @@ lazy_static! {
     );
 }
 
+pub const VERIFICATION_KEY: &[u8] = include_bytes!("data/vk-wrap-wrap-verification-key-blockchain-snark-d3623dbfa42f563e40cd5f2d032ad91f");
+
 pub struct BlockFixture {
     pub bytes: Vec<u8>,
     pub value: bin_prot::Value,
