+++ snark-keys/Cargo.toml
@@ -13,3 +13,6 @@ bin-prot = { path = "../protocol/bin-prot" }
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
 thiserror = "1"
+
+[dev-dependencies]
+test-fixtures = { path = "../protocol/test-fixtures" }
